+++ src/main/java/innexo/Period.java
@@ -2,7 +2,7 @@ package innexo;
 
 public class Period {
   public int id;
-  public int startTime;
-  public int endTime;
+  public long startTime;
+  public long endTime;
   public int period;
 }
